Added algorithm for DDR Solo Bass Mix so hardcoded decryption keys are no longer required. Refactor code so that py573a and py573a_native are one script. Native code mode can be enabled with --native now.
